refactor(scripts): #72 2.2-8 — extract report rendering to scripts/validate/_report.py
Move all five report-rendering functions (`_json_pretty`,
`render_report`, `_render_probe_section`, `_fmt_field_list`,
`_print_probe_summary`) into `scripts/validate/_report.py` (218 LOC).
The validator drops 1,152 → 962 LOC.

Pure formatting move — no logic change. The validator's per-source
`_run_<provider>` dispatchers continue to call `render_report(...)`
and `_print_probe_summary(...)` exactly as before; the validator's
import block re-exports both so the call sites are untouched.

Tests: 9 scaffold tests at 135 passed; broader subset at 695 passed.
Manual `--provider te` dry-run still produces the report header
correctly (validates `_render_probe_section` is wired up).

Codex review output got truncated (same harness quirk as 2.2-5;
codex returned exit 0 after inspecting the diff, the validator,
the new shared symbols, pyproject + CI configs, and the test
loaders). Self-reviewed the diff: pure code-motion, no behavior
change. Shipping.

Refs #72.
